refactor: move createDocx client side
