+++ functions/package.json
@@ -16,22 +16,15 @@
   },
   "main": "lib/index.js",
   "dependencies": {
-    "@local/date": "file:../shared/date",
     "@google-cloud/storage": "^5.1.1",
-    "docxtemplater": "^3.17.6",
-    "docxtemplater-image-module": "^3.1.0",
     "firebase": "^7.15.5",
     "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.6.1",
-    "pizzip": "^3.0.6",
-    "slugify": "^1.4.5"
+    "firebase-functions": "^3.6.1"
   },
   "devDependencies": {
-    "@types/pizzip": "^3.0.1",
     "dotenv": "^8.2.0",
     "firebase-functions-test": "^0.2.0",
     "mammoth": "^1.4.9",
-    "ts-node": "^8.10.2",
     "tslint": "^5.12.0",
     "typescript": "^3.8.0"
   },
