+++ package.json
@@ -22,6 +22,7 @@
     "axios": "^0.19.2",
     "date-fns": "^2.13.0",
     "docxtemplater": "^3.18.0",
+    "docxtemplater-image-module": "^3.1.0",
     "draft-js": "^0.11.6",
     "draft-js-export-html": "^1.4.1",
     "draft-js-import-html": "^1.4.1",
@@ -54,7 +55,6 @@
   "eslintConfig": {
     "extends": "react-app"
   },
-  "eslintIgnore": ["node_modules"],
   "browserslist": {
     "production": [
       ">0.2%",
