+++ types/docxtemplater-image-module/index.d.ts
@@ -1,0 +1,15 @@
+declare module "docxtemplater-image-module" {
+  export = ImageModule;
+}
+
+declare class ImageModule {
+  constructor(options?: ImageModule.Options);
+}
+
+declare namespace ImageModule {
+  interface Options {
+    centered?: boolean;
+    getImage?: any;
+    getSize?: (img: string, tagValue: string, tagName: string) => number[];
+  }
+}
